move dsmr parser to external library
