+++ esphome/components/dsmr/dsmr.h
@@ -7,16 +7,16 @@
 #include "esphome/core/log.h"
 #include "esphome/core/defines.h"
 
-#include "parser.h"
-#include "fields.h"
+#include <dsmr/parser.h>
+#include <dsmr/fields.h>
 
 namespace esphome {
-namespace dsmr_ {
+namespace dsmr {
 
 static constexpr uint32_t MAX_TELEGRAM_LENGTH = 1500;
 static constexpr uint32_t POLL_TIMEOUT = 1000;
 
-using namespace dsmr::fields;
+using namespace ::dsmr::fields;
 
 // DSMR_**_LIST generated by ESPHome and written in esphome/core/defines
 
@@ -37,12 +37,12 @@ using namespace dsmr::fields;
 #define DSMR_DATA_SENSOR(s) s
 #define COMMA ,
 
-using MyData = dsmr::ParsedData<DSMR_TEXT_SENSOR_LIST(DSMR_DATA_SENSOR, COMMA)
-                                    DSMR_BOTH DSMR_SENSOR_LIST(DSMR_DATA_SENSOR, COMMA)>;
+using MyData = ::dsmr::ParsedData<DSMR_TEXT_SENSOR_LIST(DSMR_DATA_SENSOR, COMMA)
+                                      DSMR_BOTH DSMR_SENSOR_LIST(DSMR_DATA_SENSOR, COMMA)>;
 
 class Dsmr : public Component, public uart::UARTDevice {
  public:
-  Dsmr(uart::UARTComponent* uart) : uart::UARTDevice(uart) {}
+  Dsmr(uart::UARTComponent *uart) : uart::UARTDevice(uart) {}
 
   void loop() override;
 
@@ -62,15 +62,15 @@ class Dsmr : public Component, public uart::UARTDevice {
 
   void dump_config() override;
 
-  void set_decryption_key(const std::string& decryption_key);
+  void set_decryption_key(const std::string &decryption_key);
 
 // Sensor setters
 #define DSMR_SET_SENSOR(s) \
-  void set_##s(sensor::Sensor* sensor) { s_##s##_ = sensor; }
+  void set_##s(sensor::Sensor *sensor) { s_##s##_ = sensor; }
   DSMR_SENSOR_LIST(DSMR_SET_SENSOR, )
 
 #define DSMR_SET_TEXT_SENSOR(s) \
-  void set_##s(text_sensor::TextSensor* sensor) { s_##s##_ = sensor; }
+  void set_##s(text_sensor::TextSensor *sensor) { s_##s##_ = sensor; }
   DSMR_TEXT_SENSOR_LIST(DSMR_SET_TEXT_SENSOR, )
 
  protected:
@@ -86,13 +86,13 @@ class Dsmr : public Component, public uart::UARTDevice {
   bool footer_found_{false};
 
 // Sensor member pointers
-#define DSMR_DECLARE_SENSOR(s) sensor::Sensor* s_##s##_{nullptr};
+#define DSMR_DECLARE_SENSOR(s) sensor::Sensor *s_##s##_{nullptr};
   DSMR_SENSOR_LIST(DSMR_DECLARE_SENSOR, )
 
-#define DSMR_DECLARE_TEXT_SENSOR(s) text_sensor::TextSensor* s_##s##_{nullptr};
+#define DSMR_DECLARE_TEXT_SENSOR(s) text_sensor::TextSensor *s_##s##_{nullptr};
   DSMR_TEXT_SENSOR_LIST(DSMR_DECLARE_TEXT_SENSOR, )
 
   std::vector<uint8_t> decryption_key_{};
 };
-}  // namespace dsmr_
+}  // namespace dsmr
 }  // namespace esphome
